Remove out-of-scope refactoring, simplify some logic
