+++ galley/pkg/server/settings/args.go
@@ -20,6 +20,7 @@ import (
 	"time"
 
 	"istio.io/istio/galley/pkg/crd/validation"
+	"istio.io/istio/galley/pkg/source/kube/builtin"
 	"istio.io/istio/pkg/keepalive"
 	"istio.io/istio/pkg/mcp/creds"
 	"istio.io/pkg/ctrlz"
@@ -145,7 +146,7 @@ func DefaultArgs() *Args {
 		ConfigPath:                  "",
 		DomainSuffix:                defaultDomainSuffix,
 		DisableResourceReadyCheck:   false,
-		ExcludedResourceKinds:       nil,
+		ExcludedResourceKinds:       defaultExcludedResourceKinds(),
 		SinkMeta:                    make([]string, 0),
 		KeepAlive:                   keepalive.DefaultOption(),
 		ValidationArgs:              validation.DefaultArgs(),
@@ -162,6 +163,14 @@ func DefaultArgs() *Args {
 			UpdateInterval: defaultProbeCheckInterval,
 		},
 	}
+}
+
+func defaultExcludedResourceKinds() []string {
+	resources := make([]string, 0)
+	for _, spec := range builtin.GetSchema().All() {
+		resources = append(resources, spec.Kind)
+	}
+	return resources
 }
 
 // String produces a stringified version of the arguments for debugging.
